Pequeñas mejoras al Makefile

+++ Makefile
@@ -17,11 +17,11 @@ alu: alumno.c alumno.tab.c
 
 plan_de_estudios.c: plan_de_estudios.l plan_de_estudios.tab.h
 	${LEX} $<
-	mv ${LEX_BAK} plan_de_estudios.backup
+	[ -f ${LEX_BAK} ] && mv ${LEX_BAK} plan_de_estudios.backup
 
 alumno.c: alumno.l alumno.tab.h
 	${LEX} $<
-	mv ${LEX_BAK} alumno.backup
+	[ -f ${LEX_BAK} ] && mv ${LEX_BAK} alumno.backup
 
 %.tab.h: %.y
 	${YACC} $<
@@ -30,7 +30,6 @@ tags: alumno.l alumno.y plan_de_estudios.l plan_de_estudios.y funciones_plan.h e
 	${TAGS} ${TAGS_OPT} $^
 
 clean:
-	@echo "Limpiando..."
-	rm -f *.tab.* *.c *.backup pe alu tags *.output
+	rm -f *.tab.[ch] *.c *.backup pe alu *.output
 
 .PHONY: all clean
